libs/policy-change-request/src/lib/policy-change-request-flow/policy-transactions/remove-dependent/remove-dependent.component.ts
import { Component, OnInit, OnDestroy } from "@angular/core";
import { FormBuilder, FormGroup, Validators, FormArray } from "@angular/forms";
import { MatDialog, MatDialogRef } from "@angular/material/dialog";
import {
    PolicyChangeRequestList,
    AppSettings,
    ReasonToRemoveDependant,
    Relationship,
    Title,
    CoverageLevels,
    CountryState,
} from "@empowered/constants";
import { Observable, Subscription, forkJoin } from "rxjs";
import { StaticService, MemberService, AccountService, PolicyTransactionForms } from "@empowered/api";
import { LanguageService } from "@empowered/language";
import { PolicyChangeRequestState, SetRemoveDependantRequest, SharedState } from "@empowered/ngxs-store";
import { Select, Store } from "@ngxs/store";
import { mergeMap } from "rxjs/operators";
import { SideNavService } from "../../side-nav/services/side-nav.service";
import { DatePipe } from "@angular/common";
import { HttpEvent } from "@angular/common/http";
import { NgxMaskPipe } from "ngx-mask";
import { PolicyChangeRequestComponent } from "../../../policy-change-request.component";
import { ReasonRemovalDependent } from "@empowered/constants";
import { PolicyChangeRequestCancelPopupComponent, PolicyChangeRequestConfirmationPopupComponent } from "@empowered/ui";

@Component({
    selector: "empowered-remove-dependent",
    templateUrl: "./remove-dependent.component.html",
    styleUrls: ["./remove-dependent.component.scss"],
    providers: [DatePipe],
})
export class RemoveDependentComponent implements OnInit, OnDestroy {
    removeDependantForm: FormGroup;
    validationRegex: any;
    subscriptions: Subscription[] = [];
    suffixes$: Observable<string[]>;
    gender$: Observable<string[]>;
    states$: Observable<CountryState[]>;
    policyList = [];
    mpGroup: number;
    memberId: number;
    planId: any;
    showSpinner: boolean;
    dependantList = [];
    other = {
        relationship: "Other",
        birthDate: null,
        gender: null,
        id: "Other",
    };
    isOtherSelected: boolean;
    counter = 0;
    isAllPolicySelected: boolean;
    coverageLevel = Object.values(CoverageLevels);
    reasonList: Array<ReasonRemovalDependent>;
    titleList = Object.values(Title);
    relationshipList = [];
    isSupportiveDocumentsRequired: boolean;
    dependantRelations = [];
    isSubmitted: boolean;
    uploadApi: Observable<HttpEvent<any>>;
    isIndeterminate: boolean;
    STR_CHILD = "CHILD";
    STR_DIVORCE = "Divorce";
    STR_SPOUSE = "SPOUSE";
    STR_DEPENDENT_ATTAINING_AGE = "DependentAttainingAge";
    languageStrings: Record<string, string> = this.languageService.fetchPrimaryLanguageValues([
        "primary.portal.policyChangeRequest.transactions.removeDependant.header",
        "primary.portal.policyChangeRequest.transactions.continue",
        "primary.portal.policyChangeRequest.transactions.cancel",
        "primary.portal.policyChangeRequest.transactions.removeDependant.otherDependantTooltip",
        "primary.portal.dashboard.policyChangeRequestFlow.pcrFlow",
        "primary.portal.common.cancel",
        "primary.portal.common.back",
        "primary.portal.dashboard.policyChangeRequestFlow.cancelMessage",
        "primary.portal.dashboard.policyChangeRequestFlow.cancelMessageNotSaved",
        "primary.portal.policyChangeRequest.transactions.back",
        "primary.portal.policyChangeRequest.transactions.removeDependant.affectedPolicy",
        "primary.portal.policyChangeRequest.transactions.selectAll",
        "primary.portal.policyChangeRequest.transactions.removeDependant.dependantToRemoved",
        "primary.portal.policyChangeRequest.transactions.requiredField",
        "primary.portal.policyChangeRequest.transactions.changeName.title",
        "primary.portal.policyChangeRequest.transactions.removeDependant.billingName",
        "primary.portal.removeDependentReason.death",
        "primary.portal.removeDependentReason.dependentAttainingAge",
        "primary.portal.removeDependentReason.divorce",
        "primary.portal.removeDependentReason.request",
        "primary.portal.common.invalidDate",
    ]);
    removeDependantRequestInitialData: any;
    documentIdArray = [];
    cifNumber: any;
    minDate = new Date();
    maxDate = null;
    nameWithHypenApostrophesValidation: any;
    originalReasonList: ReasonRemovalDependent[];
    selectedPolicyIds = [];
    formArray: any;
    storePolicy: any;
    memberData: any;
    dependentId: number;
    @Select(PolicyChangeRequestState.GetmemberInfo) memberInfo$: Observable<any>;
    @Select(PolicyChangeRequestState.GetPlanId) planId$: Observable<string>;
    @Select(SharedState.regex) regex$: Observable<any>;
    @Select(PolicyChangeRequestState.GetRemoveDependantRequest) removeDependantRequest$: Observable<any>;

    constructor(
        private readonly fb: FormBuilder,
        private readonly dialog: MatDialog,
        private readonly staticService: StaticService,
        private readonly languageService: LanguageService,
        private readonly memberService: MemberService,
        private readonly accountService: AccountService,
        private readonly store: Store,
        private readonly sideNavService: SideNavService,
        private readonly datePipe: DatePipe,
        private readonly maskPipe: NgxMaskPipe,
        private cancelDialogRef: MatDialogRef<PolicyChangeRequestCancelPopupComponent>,
        private readonly PCRDialogRef: MatDialogRef<PolicyChangeRequestComponent>,
    ) {
        this.subscriptions.push(
            this.regex$.subscribe((data) => {
                if (data) {
                    this.validationRegex = data;
                }
            }),
        );
    }
    /**
     * @description function will initialize necessary variable and configure data and form needed for component
     * @returns void
     */
    ngOnInit(): void {
        this.originalReasonList = [
            { key: "Death", value: this.languageStrings["primary.portal.removeDependentReason.death"] },
            {
                key: "DependentAttainingAge",
                value: this.languageStrings["primary.portal.removeDependentReason.dependentAttainingAge"],
            },
            { key: "Divorce", value: this.languageStrings["primary.portal.removeDependentReason.divorce"] },
            { key: "Request", value: this.languageStrings["primary.portal.removeDependentReason.request"] },
        ];
        this.reasonList = [...this.originalReasonList];
        this.relationshipList = this.convertEnumToKeyValuePair(Relationship);
        this.nameWithHypenApostrophesValidation = new RegExp(this.validationRegex.NAME_WITH_HYPENS_APOSTROPHES);
        this.removeDependantForm = this.fb.group(
            {
                policyNumbers: this.fb.array([]),
                dependentId: ["", Validators.required],
                reasonForRemoval: ["", Validators.required],
                effectiveDate: [new Date(), Validators.required],
                documentIds: [[], Validators.required],
                newCoverageLevelName: ["", Validators.required],
                billingName: this.fb.group({
                    firstName: ["", Validators.compose([Validators.required, Validators.pattern(new RegExp(this.validationRegex.NAME))])],
                    middleName: [
                        "",
                        Validators.compose([Validators.maxLength(60), Validators.pattern(new RegExp(this.validationRegex.NAME))]),
                    ],
                    lastName: ["", Validators.compose([Validators.required, Validators.pattern(new RegExp(this.validationRegex.NAME))])],
                    suffix: [""],
                }),
                type: Object.keys(PolicyTransactionForms)[4],
            },
            { updateOn: "blur" },
        );
        this.removeDependantRequestInitialData = { ...this.removeDependantForm.value };

        if (this.store.selectSnapshot(PolicyChangeRequestState.GetmemberInfo)) {
            this.memberData = this.store.selectSnapshot(PolicyChangeRequestState.GetmemberInfo);
            this.policyList = this.memberData["policies"];
            this.mpGroup = this.memberData["groupId"];
            this.memberId = this.memberData["memberId"];
            this.cifNumber = this.memberData["cifNumber"];
        }

        this.subscriptions.push(
            this.removeDependantRequest$.subscribe((removeDependantRequest) => {
                if (removeDependantRequest) {
                    this.dependentId = removeDependantRequest.dependentId;
                    if (removeDependantRequest.reasonForRemoval === ReasonToRemoveDependant.DEATH) {
                        this.minDate = null;
                        this.maxDate = new Date();
                    }
                    this.getDataFromStore(removeDependantRequest);
                    this.loadDropDowns();
                } else {
                    this.getMemberInfo();
                }
            }),
        );
    }

    getDataFromStore(removeDependantRequest: any): void {
        const dependentRequest = { ...removeDependantRequest };
        if (removeDependantRequest["otherDependentName"]) {
            dependentRequest["dependentId"] = "Other";
            this.addOtherFormControl();
        }
        this.removeDependantRequestInitialData = {
            ...this.store.selectSnapshot(PolicyChangeRequestState.GetRemoveDependantInitialData),
        };
        this.counter = removeDependantRequest["policyNumbers"].length;
        this.documentIdArray = removeDependantRequest.documentIds;
        this.storePolicy = [...removeDependantRequest["policyNumbers"]];
        this.convertAffectedPolicy();
        this.removeDependantForm.patchValue(dependentRequest);
        this.checkForOtherDependant(removeDependantRequest);
        this.createPolicyFormContol(true);
        this.checkForIntermidiateValue();
    }

    convertAffectedPolicy(): void {
        if (this.storePolicy.length) {
            this.policyList.forEach((policy) => {
                const matchedPolicy = this.storePolicy.find((policyNumber) => policyNumber === policy.policyNumber);
                if (matchedPolicy) {
                    this.storePolicy[this.storePolicy.indexOf(matchedPolicy)] = policy;
                }
            });
        }
    }

    convertEnumToKeyValuePair(data: any): any {
        return Object.keys(data).map((key) => ({ id: key, name: data[key] }));
    }

    checkForOtherDependant(removeDependantRequest: any): void {
        if (removeDependantRequest.other) {
            this.addOtherFormControl();
            this.removeDependantForm.patchValue({
                other: removeDependantRequest.other,
            });
        }
    }

    cancel(): void {
        this.cancelDialogRef = this.dialog.open(PolicyChangeRequestCancelPopupComponent, {
            width: "667px",
            data: {
                cancelModalDisplayType: this.languageStrings["primary.portal.dashboard.policyChangeRequestFlow.pcrFlow"],
                cancelButton: this.languageStrings["primary.portal.common.cancel"],
                backButton: this.languageStrings["primary.portal.common.back"],
                requestType: this.languageStrings["primary.portal.dashboard.policyChangeRequestFlow.cancelMessage"],
                description: this.languageStrings["primary.portal.dashboard.policyChangeRequestFlow.cancelMessageNotSaved"],
            },
        });

        this.subscriptions.push(
            this.cancelDialogRef.afterClosed().subscribe((result) => {
                if (result === AppSettings.CANCEL) {
                    this.sideNavService.removeTransactionScreenFromStore(true);
                    this.store.dispatch(new SetRemoveDependantRequest(null, null));
                    this.PCRDialogRef.close(PolicyChangeRequestList.cancel);
                }
            }),
        );
    }

    back(): void {
        this.sideNavService.onBackClick();
    }

    get billingNameControl(): unknown {
        return this.formControl["billingName"].controls;
    }

    checkDependantList(): void {
        if (this.dependantList.length > 0) {
            this.removeDependantForm.patchValue({
                dependentId: "Other",
            });
            this.addOtherFormControl();
        }
    }
    /**
     * Function to load form dropdown data
     */
    loadDropDowns(): void {
        this.suffixes$ = this.staticService.getSuffixes();
        this.gender$ = this.staticService.getGenders();
        this.states$ = this.staticService.getStates();
        this.dependantList.push(this.other);
        if (this.memberId && this.mpGroup) {
            this.showSpinner = true;
            this.subscriptions.push(
                forkJoin([
                    this.memberService.getMemberDependents(this.memberId, false, this.mpGroup),
                    this.accountService.getDependentRelations(this.mpGroup),
                ]).subscribe(
                    (result) => {
                        this.showSpinner = false;
                        this.dependantList = this.dependantList.concat(result[0]);
                        this.dependantRelations = result[1];
                        if (this.dependentId) {
                            const dependentValue = this.dependantList.find((dependent) => dependent.id === this.dependentId);
                            if (dependentValue) {
                                const relation = this.dependantRelations.find(
                                    (dependantRelations) => dependantRelations.id === dependentValue.dependentRelationId,
                                );
                                if (relation) {
                                    this.removeReasonBasedOnSelectedDependant(relation.code);
                                }
                            }
                        }
                        this.mapDependantWithRelationship();
                    },
                    (error) => {
                        this.showSpinner = false;
                    },
                ),
            );
        }
    }
    /**
     * @description Functions will be triggered when relation to member is changed and will populate reason
     *  of removal depending upon relation to member
     * @param relation: relation to member
     * @returns void
     */
    removeReasonBasedOnSelectedDependant(relation: string): void {
        this.reasonList = [...this.originalReasonList];
        if (relation === this.STR_CHILD) {
            this.reasonList.splice(
                this.reasonList.findIndex((reason) => reason.key === this.STR_DIVORCE),
                1,
            );
        } else if (relation === this.STR_SPOUSE) {
            this.reasonList.splice(
                this.reasonList.findIndex((reason) => reason.key === this.STR_DEPENDENT_ATTAINING_AGE),
                1,
            );
        }
    }

    mapDependantWithRelationship(): void {
        if (this.dependantList) {
            this.dependantList.forEach((dependant) => {
                const dependantRelation = this.dependantRelations.find((relation) => relation["id"] === dependant["dependentRelationId"]);
                if (dependantRelation) {
                    dependant["relationship"] = dependantRelation["relationType"];
                }
            });
        }
        this.showSpinner = false;
    }

    getMemberInfo(): void {
        this.createPolicyFormContol(false);
        this.loadDropDowns();
        this.getMemeberData();
    }

    getMemeberData(): void {
        if (this.mpGroup && this.memberId) {
            this.showSpinner = true;
            this.subscriptions.push(
                this.accountService
                    .getAccount(this.mpGroup.toString())
                    .pipe(
                        mergeMap((response) => {
                            if (
                                response["partnerAccountType"] === AppSettings.PAYROLL_ACCOUNT ||
                                response["partnerAccountType"] === AppSettings.ASSOCIATION_ACCOUNT
                            ) {
                                return this.memberService.getMember(this.memberId, false, this.mpGroup.toString());
                            }
                            return undefined;
                        }),
                    )
                    .subscribe(
                        (data) => {
                            this.removeDependantForm.patchValue({
                                billingName: data["body"].name,
                            });
                            this.showSpinner = false;
                        },
                        (error) => {
                            this.showSpinner = false;
                        },
                    ),
            );
        }
    }

    createPolicyFormContol(selected?: boolean): void {
        this.isAllPolicySelected = selected;
        this.formArray = this.removeDependantForm.get("policyNumbers") as FormArray;
        if (this.storePolicy) {
            this.policyList.forEach((element) => {
                const setValue = this.storePolicy.indexOf(element) !== -1;
                this.formArray.push(this.fb.control(setValue));
            });
        } else {
            this.policyList.forEach((x) => this.formArray.push(this.fb.control(false)));
        }
    }

    selectAll(isChecked: boolean): void {
        this.formControl["policyNumbers"].controls = [];
        if (isChecked) {
            this.counter = this.policyList.length;
            this.isAllPolicySelected = true;
            this.policyList.forEach((x) => this.formArray.push(this.fb.control(true)));
        } else {
            this.isAllPolicySelected = false;
            this.counter = 0;
            this.isIndeterminate = false;
            this.policyList.forEach((x) => this.formArray.push(this.fb.control(false)));
        }
    }

    checkAllpolicySelected(): void {
        if (this.counter > 0) {
            this.counter--;
        }
    }

    selectSingle(isChecked: boolean): void {
        if (!isChecked) {
            this.counter = this.counter > 0 ? --this.counter : this.counter;
            this.isAllPolicySelected = isChecked;
        } else {
            this.counter++;
            this.isAllPolicySelected = this.counter === this.formControl["policyNumbers"].controls.length ? true : false;
        }
        this.checkForIntermidiateValue();
    }

    checkForIntermidiateValue(): void {
        this.isAllPolicySelected = this.counter === this.policyList.length ? true : false;
        this.isIndeterminate = this.counter !== 0 && !this.isAllPolicySelected ? true : false;
    }
    /**
     * Get form controls
     */
    get formControl(): unknown {
        return this.removeDependantForm.controls;
    }

    /**
     * Submit transfer to direct request
     */
    removeDependent(): void {
        if (!this.removeDependantForm.dirty && !this.store.selectSnapshot(PolicyChangeRequestState.GetRemoveDependantRequest)) {
            this.openConfirmationPopup();
        } else {
            this.removeDependantForm.value["documentIds"] = this.documentIdArray;
            this.isSubmitted = true;
            this.validateDocument();
            this.validateAllFormFields(this.removeDependantForm);
            this.setAffectedPolicy();
            this.storeRequest();
        }
    }
    setAffectedPolicy(): void {
        this.selectedPolicyIds = this.policyList.filter((x, i) => !!this.removeDependantForm.value.policyNumbers[i]);
        this.selectedPolicyIds = this.selectedPolicyIds.map((policy) => policy.policyNumber);
    }

    storeRequest(): void {
        this.removeDependantForm.value["documentIds"] = this.documentIdArray;
        this.removeDependantForm.value["policyNumbers"] = this.selectedPolicyIds;
        this.counter = this.selectedPolicyIds.length;
        if (this.removeDependantForm.valid && this.counter) {
            this.setDateFormat();
            if (this.removeDependantForm.value["dependentId"] === "Other") {
                delete this.removeDependantForm.value["dependentId"];
            }
            this.store.dispatch(new SetRemoveDependantRequest(this.removeDependantForm.value, this.removeDependantRequestInitialData));
            this.sideNavService.onNextClick(1);
        }
    }

    setDateFormat(): void {
        this.removeDependantForm.value["effectiveDate"] = this.datePipe.transform(
            this.removeDependantForm.value["effectiveDate"],
            AppSettings.DATE_FORMAT,
        );
    }

    validateAllFormFields(formGroup: FormGroup): void {
        Object.keys(formGroup.controls).forEach((field) => {
            const control = formGroup.get(field);
            if (control["controls"]) {
                for (const subField in control["controls"]) {
                    if (subField) {
                        control["controls"][subField].markAsTouched({ onlySelf: true });
                    }
                }
            } else {
                control.markAsTouched({ onlySelf: true });
            }
        });
    }

    validateDocument(): void {
        if (this.removeDependantForm.value["documentIds"].length === 0) {
            this.isSupportiveDocumentsRequired = true;
        } else {
            this.isSupportiveDocumentsRequired = false;
            this.formControl["documentIds"].clearValidators();
            this.formControl["documentIds"].updateValueAndValidity();
        }
    }

    getDocumentId(documentID: number): void {
        if (documentID) {
            this.documentIdArray.push(documentID);
            this.isSupportiveDocumentsRequired = false;
            this.formControl["documentIds"].clearValidators();
            this.formControl["documentIds"].updateValueAndValidity();
        } else {
            this.documentIdArray = [];
            this.isSupportiveDocumentsRequired = false;
            this.formControl["documentIds"].setValidators([Validators.required]);
            this.formControl["documentIds"].updateValueAndValidity();
        }
    }
    /**
     * Open confirmation popup
     */
    openConfirmationPopup(): void {
        const dialogRef = this.dialog.open(PolicyChangeRequestConfirmationPopupComponent, {
            width: "667px",
            data: {
                cancelButton: this.languageStrings["primary.portal.policyChangeRequest.transactions.cancel"],
                continueButton: this.languageStrings["primary.portal.policyChangeRequest.transactions.continue"],
                requestType: this.languageStrings["primary.portal.policyChangeRequest.transactions.removeDependant.header"],
            },
        });
        this.subscriptions.push(
            dialogRef.afterClosed().subscribe((result) => {
                if (result === AppSettings.CONTINUE) {
                    this.sideNavService.onNextClick(1);
                } else {
                    dialogRef.close();
                }
            }),
        );
    }

    /**
     * Add other control on selection of other coverage level other
     */
    addOtherFormControl(): void {
        const otherDependentGender = this.fb.control("", [Validators.required]);
        const otherDependentRelationship = this.fb.control("", [Validators.required]);
        const otherDependentName = this.fb.group({
            title: [""],
            firstName: ["", Validators.compose([Validators.required, Validators.pattern(new RegExp(this.validationRegex.NAME))])],
            middleName: ["", Validators.compose([Validators.pattern(new RegExp(this.validationRegex.NAME))])],
            lastName: ["", Validators.required],
            suffix: [""],
        });
        const otherDependentAddress = this.fb.group({
            address1: [
                "",
                Validators.compose([
                    Validators.required,
                    Validators.maxLength(100),
                    Validators.pattern(new RegExp(this.validationRegex.ADDRESS)),
                ]),
            ],
            address2: ["", Validators.compose([Validators.maxLength(100), Validators.pattern(new RegExp(this.validationRegex.ADDRESS))])],
            city: [
                "",
                Validators.compose([
                    Validators.required,
                    Validators.maxLength(100),
                    Validators.pattern(new RegExp(this.validationRegex.ADDRESS)),
                ]),
            ],
            state: ["", Validators.required],
            zip: ["", Validators.compose([Validators.required, Validators.pattern(new RegExp(this.validationRegex.ZIP_CODE))])],
        });
        this.removeDependantForm.addControl("otherDependentAddress", otherDependentAddress);
        this.removeDependantForm.addControl("otherDependentRelationship", otherDependentRelationship);
        this.removeDependantForm.addControl("otherDependentGender", otherDependentGender);
        this.removeDependantForm.addControl("otherDependentName", otherDependentName);
        this.removeDependantRequestInitialData = {
            ...this.removeDependantRequestInitialData,
            otherDependentAddress: otherDependentAddress.value,
            otherDependentRelationship: otherDependentRelationship.value,
            otherDependentGender: otherDependentGender.value,
            otherDependentName: otherDependentName.value,
        };
    }

    get otherDependentNameFormControl(): void {
        return this.formControl["otherDependentName"].controls;
    }

    get otherDependentAddressFromControl(): void {
        return this.formControl["otherDependentAddress"].controls;
    }

    /**
     * Add form control on selction other dependant
     * @param dependant
     * @param isOtherChecked
     */
    otherDependantSelected(dependantSelected: string, relation?: string): void {
        this.removeReasonBasedOnSelectedDependant(relation);
        if (dependantSelected === "Other") {
            this.addOtherFormControl();
        } else {
            this.removeControls();
        }
    }

    removeControls(): void {
        this.removeDependantForm.removeControl("otherDependentAddress");
        this.removeDependantForm.removeControl("otherDependentRelationship");
        this.removeDependantForm.removeControl("otherDependentGender");
        this.removeDependantForm.removeControl("otherDependentName");
        delete this.removeDependantRequestInitialData.otherDependentAddress;
        delete this.removeDependantRequestInitialData.otherDependentRelationship;
        delete this.removeDependantRequestInitialData.otherDependentGender;
        delete this.removeDependantRequestInitialData.otherDependentName;
    }

    /**
     * Function to validate Zip Code
     */
    validateZipCode(): void {
        if (
            this.otherDependentAddressFromControl["zip"].dirty &&
            this.otherDependentAddressFromControl["state"].value &&
            this.otherDependentAddressFromControl["zip"].value
        ) {
            this.showSpinner = true;
            this.subscriptions.push(
                this.staticService
                    .validateStateZip(
                        this.otherDependentAddressFromControl["state"].value,
                        this.otherDependentAddressFromControl["zip"].value,
                    )
                    .subscribe(
                        (resp) => {
                            this.showSpinner = false;
                            this.otherDependentAddressFromControl["zip"].setErrors(null);
                            this.otherDependentAddressFromControl["state"].setErrors(null);
                        },
                        (error) => {
                            this.showSpinner = false;
                            this.otherDependentAddressFromControl["zip"].setErrors({ zipValid: true });
                            this.otherDependentAddressFromControl["state"].setErrors({
                                zipValid: true,
                            });
                        },
                    ),
            );
        }
    }
    /**
     * This method will unsubscribe all the api subscription.
     */

    transform(event: any): void {
        event.target.value = this.maskPipe.transform(event.target.value, AppSettings.DATE_MASK_FORMAT);
    }

    /**
     * Function to set date of event based on removal selection
     * @removalType is selected removal type
     */
    reasonForRemovalSelection(removalType: string): void {
        this.minDate = new Date();
        this.maxDate = null;
        if (removalType === ReasonToRemoveDependant.DEATH) {
            this.minDate = null;
            this.maxDate = new Date();
        }
    }

    ngOnDestroy(): void {
        if (this.subscriptions && this.subscriptions.length > 0) {
            this.subscriptions.forEach((sub) => sub.unsubscribe());
        }
    }
}
